Add parsing logic for Snyk API errors
Also move tests for SnykParser into their own class instead of keeping them in SnykAnalysisTaskTest.
